MC-36: Add arrive-by state to RouteInputForm story
