+++ src/components/RouteInputForm/RouteInputForm.jsx
@@ -7,7 +7,7 @@ const RouteInputForm = ({
   onStartingPointChange,
   onDestinationChange,
   onArriveByDateTimeChange,
-  arriveByValue,
+  arriveByDateTimeValue,
 }) => {
   return (
     <form>
@@ -28,7 +28,7 @@ const RouteInputForm = ({
       <div>
         <label id="arrive-by-label">Arrive By</label>
         <DateTimeInput
-          value={arriveByValue}
+          value={arriveByDateTimeValue}
           onChange={onArriveByDateTimeChange}
         />
       </div>
@@ -40,7 +40,7 @@ RouteInputForm.propTypes = {
   onStartingPointChange: PropTypes.func.isRequired,
   onDestinationChange: PropTypes.func.isRequired,
   onArriveByDateTimeChange: PropTypes.func.isRequired,
-  arriveByValue: PropTypes.string.isRequired,
+  arriveByDateTimeValue: PropTypes.string.isRequired,
 };
 
 export default RouteInputForm;
